* MISUV-9397: ** - re-design extensions for FinancialDetailsModel;
